requirements.txt
ccxt
pandas
numpy<2.0
pandas-ta
pyyaml
python-dotenv
pytest
